use default config when configmap cannot be found
currently, if the config for either the operator or each provider
cannot be found, the operator will crash with a config missing error.

this is not ideal, and with rhmi we would like to not have to
pre-create configmaps.

this solution, simply provides an entirely default configmap that
is used if the expected configmap is not found.