prefix s3 bucket names with cluster identifier
currently, when s3 buckets and other aws resources are created,
their names are deterministic and based off of the name of the
custom resource which triggers their creation.

this change ensures that a cluster identifier is used in front of
s3 bucket names. so that resources created using a cr in a
namespace with identical names in different clusters will not
clash