add provider specified reconcile timeouts
currently we have a global override for the reconcile timings in
each resource type. however this should be more individual instead
of allowing a controller to decide what the timings should be.

this change performs a slight refactor to the current reconcile
time override function and also ensures each provider can return
it's own reconcile time based on knowledge of the cr state.