fix region reconciliation in aws service providers
currently aws regions are not being set in a way which would
benefit rhmi, instead defaulting to eu-west-1 for aws deploys.

this change sets a priority to different ways of retrieving the
aws region in aws providers:
- try from the configmap strategy 'region' key
- try from the 'cluster' infrastructure cr in openshift
- error

this results in there being no more hardcoded region in the
operator, which was used heavily during development. and ensures
that, unless explicitly overridden with a configmap, rhmi services
will attempt to be provisioned in the same region as the openshift
cluster that the operator is running in.

verification:
- run the operator against a non-eu-west-1 region cluster
- ensure services come up as expected