remove hardcoded regions from the aws strategy configmap example
currently the example aws strategy configmap is created in dev make
targets. this means that the hardcoded region of eu-west-1 is used
even if the cluster isn't in eu-west-1 which can cause issues with
vpc setup.

remove the hardcoded region but keep the explicit keys to let users
know the possible keys that can be used in the strategy configmap.

verification:
- in cluster in a non-eu-west-1 region
- run make cluster/prepare NAMESPACE=<your-namespace>
- ensure the region isn't specified in the aws strategy configmap
- create a postgres, redis and blobstorage and ensure each are
  created in the region of the cluster